requirements.txt
attrs==21.4.0
black==22.1.0
boto3==1.24.22
botocore==1.27.22
click==8.0.4
contourpy==1.0.6
coverage==6.4.1
cycler==0.11.0
digart @ file:///Users/robertreilly/code/digital_art
exif==1.3.5
Flask==2.1.2
fonttools==4.38.0
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mypy-extensions==0.4.3
numpy==1.22.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.5.1
plotly==5.5.0
pluggy==1.0.0
plum-py==0.8.5
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
